+++ src/Data/Ord/Deriving/Internal.hs
@@ -452,7 +452,7 @@ makeOrdFunForType oFun tvMap conName ty = do
         tyVarNames :: [Name]
         tyVarNames = Map.keys tvMap
 
-    itf <- isTyFamily tyCon
+    itf <- isInTypeFamilyApp tyVarNames tyCon tyArgs
     if any (`mentionsName` tyVarNames) lhsArgs
           || itf && any (`mentionsName` tyVarNames) tyArgs
        then outOfPlaceTyVarError oClass conName