+++ src/Text/Read/Deriving/Internal.hs
@@ -687,7 +687,7 @@ makeReadForType rClass urp tvMap conName tyExpName rl ty = do
         tyVarNames :: [Name]
         tyVarNames = Map.keys tvMap
 
-    itf <- isTyFamily tyCon
+    itf <- isInTypeFamilyApp tyVarNames tyCon tyArgs
     if any (`mentionsName` tyVarNames) lhsArgs
           || itf && any (`mentionsName` tyVarNames) tyArgs
        then outOfPlaceTyVarError rClass conName