+++ src/Text/Show/Deriving/Internal.hs
@@ -555,7 +555,7 @@ makeShowForType sClass conName tvMap sl ty = do
         tyVarNames :: [Name]
         tyVarNames = Map.keys tvMap
 
-    itf <- isTyFamily tyCon
+    itf <- isInTypeFamilyApp tyVarNames tyCon tyArgs
     if any (`mentionsName` tyVarNames) lhsArgs
           || itf && any (`mentionsName` tyVarNames) tyArgs
        then outOfPlaceTyVarError sClass conName